requirements.txt
Flask==1.0.3
python-dotenv==0.10.3
Flask-PyMongo==2.3.0